create_key/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
pycryptodome==3.19.0
PyMySQL==1.1.0
requests==2.28.2
SQLAlchemy==2.0.23
typing_extensions==4.8.0
python-dotenv==1.0.0
